requirements.txt
# Core Libraries
litellm
python-dotenv>=1.0.0
PyYAML==6.0
tiktoken
openai
anthropic
google-generativeai
gradio
docker
Jinja2
pydantic
watchdog
termcolor
requests
typing-extensions

# Linear Integration Dependencies
graphql-core>=3.2.0
aiohttp>=3.8.0
cryptography>=3.4.0

# LLM Models & Frameworks
# transformers
# torch
# accelerate
# bitsandbytes # For quantization if needed
# sentencepiece

# Evolutionary Algorithms / RL
# deap
# pygad
# ray[rllib]
# stable-baselines3

# Evaluation & Sandboxing
pytest
# hypothesis
# e2b-sdk  # For E2B sandbox

# Data Storage
# psycopg2-binary  # For PostgreSQL
# pymongo  # For MongoDB
# dvc

# Utilities
# jinja2

# API/Web (Optional, if exposing functionality)
# fastapi
# uvicorn

# For potential future use or specific agent needs (example)
